PCA

PCA (Principal Component Analysis) is an unsupervised dimensionality reduction technique used in data analysis and machine learning. It transforms a dataset of possibly correlated variables into a new set of uncorrelated variables called principal components.

Usage

dt_pca(attribute = NULL, components = NULL)

Arguments

attribute

target attribute to model building

components

number of components for PCA

Value

returns an object of class dt_pca

Examples

mypca <- dt_pca("Species")
# Automatically fitting number of components
mypca <- fit(mypca, iris)
iris.pca <- transform(mypca, iris)
head(iris.pca)
#>        PC1       PC2 Species
#> 1 2.640270 -5.204041  setosa
#> 2 2.670730 -4.666910  setosa
#> 3 2.454606 -4.773636  setosa
#> 4 2.545517 -4.648463  setosa
#> 5 2.561228 -5.258629  setosa
#> 6 2.975946 -5.707321  setosa
head(mypca$pca.transf)
#>                     PC1         PC2
#> Sepal.Length  0.5210659 -0.37741762
#> Sepal.Width  -0.2693474 -0.92329566
#> Petal.Length  0.5804131 -0.02449161
#> Petal.Width   0.5648565 -0.06694199
# Manual establishment of number of components
mypca <- dt_pca("Species", 3)
mypca <- fit(mypca, datasets::iris)
iris.pca <- transform(mypca, iris)
head(iris.pca)
#>        PC1       PC2      PC3 Species
#> 1 2.640270 -5.204041 2.488621  setosa
#> 2 2.670730 -4.666910 2.466898  setosa
#> 3 2.454606 -4.773636 2.288321  setosa
#> 4 2.545517 -4.648463 2.212378  setosa
#> 5 2.561228 -5.258629 2.392226  setosa
#> 6 2.975946 -5.707321 2.437245  setosa
head(mypca$pca.transf)
#>                     PC1         PC2        PC3
#> Sepal.Length  0.5210659 -0.37741762  0.7195664
#> Sepal.Width  -0.2693474 -0.92329566 -0.2443818
#> Petal.Length  0.5804131 -0.02449161 -0.1421264
#> Petal.Width   0.5648565 -0.06694199 -0.6342727