Smoothing by interval

The "smoothing by interval" function is used to apply a smoothing technique to a vector or time series data using a moving window approach.

Usage

smoothing_inter(n)

Arguments

n

number of bins

Value

returns an object of class smoothing_inter

Examples

data(iris)
obj <- smoothing_inter(n = 2)
obj <- fit(obj, iris$Sepal.Length)
sl.bi <- transform(obj, iris$Sepal.Length)
table(sl.bi)
#> sl.bi
#> 5.32842105263158 6.73272727272727 
#>               95               55 
obj$interval
#> [1] 4.3 6.1 7.9

entro <- evaluate(obj, as.factor(names(sl.bi)), iris$Species)
entro$entropy
#> [1] 1.191734