Subset Extraction for Time Series Data

Extracts a subset of a time series object based on specified rows and columns. The function allows for flexible indexing and subsetting of time series data.

Usage

# S3 method for class 'ts_data'
x[i, j, ...]

Arguments

x

ts_data object

i

row i

j

column j

...

optional arguments

Value

returns a new ts_data object

Examples

data(sin_data)
data10 <- ts_data(sin_data$y, 10)
ts_head(data10)
#>             t9        t8        t7        t6        t5        t4        t3
#> [1,] 0.0000000 0.2474040 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950
#> [2,] 0.2474040 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859
#> [3,] 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859 0.9092974
#> [4,] 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859 0.9092974 0.7780732
#> [5,] 0.8414710 0.9489846 0.9974950 0.9839859 0.9092974 0.7780732 0.5984721
#> [6,] 0.9489846 0.9974950 0.9839859 0.9092974 0.7780732 0.5984721 0.3816610
#>             t2         t1         t0
#> [1,] 0.9839859  0.9092974  0.7780732
#> [2,] 0.9092974  0.7780732  0.5984721
#> [3,] 0.7780732  0.5984721  0.3816610
#> [4,] 0.5984721  0.3816610  0.1411200
#> [5,] 0.3816610  0.1411200 -0.1081951
#> [6,] 0.1411200 -0.1081951 -0.3507832
#single line
data10[12,]
#>            t9      t8         t7         t6         t5         t4         t3
#> [1,] 0.381661 0.14112 -0.1081951 -0.3507832 -0.5715613 -0.7568025 -0.8949894
#>              t2         t1         t0
#> [1,] -0.9775301 -0.9992928 -0.9589243
#> attr(,"class")
#> [1] "ts_data" "matrix"  "array"  
#> attr(,"sw")
#> [1] 10

#range of lines
data10[12:13,]
#>            t9         t8         t7         t6         t5         t4         t3
#> [1,] 0.381661  0.1411200 -0.1081951 -0.3507832 -0.5715613 -0.7568025 -0.8949894
#> [2,] 0.141120 -0.1081951 -0.3507832 -0.5715613 -0.7568025 -0.8949894 -0.9775301
#>              t2         t1         t0
#> [1,] -0.9775301 -0.9992928 -0.9589243
#> [2,] -0.9992928 -0.9589243 -0.8589345
#> attr(,"class")
#> [1] "ts_data" "matrix"  "array"  
#> attr(,"sw")
#> [1] 10

#single column
data10[,1]
#>                t9
#>  [1,]  0.00000000
#>  [2,]  0.24740396
#>  [3,]  0.47942554
#>  [4,]  0.68163876
#>  [5,]  0.84147098
#>  [6,]  0.94898462
#>  [7,]  0.99749499
#>  [8,]  0.98398595
#>  [9,]  0.90929743
#> [10,]  0.77807320
#> [11,]  0.59847214
#> [12,]  0.38166099
#> [13,]  0.14112001
#> [14,] -0.10819513
#> [15,] -0.35078323
#> [16,] -0.57156132
#> [17,] -0.75680250
#> [18,] -0.89498936
#> [19,] -0.97753012
#> [20,] -0.99929279
#> [21,] -0.95892427
#> [22,] -0.85893449
#> [23,] -0.70554033
#> [24,] -0.50827908
#> [25,] -0.27941550
#> [26,] -0.03317922
#> [27,]  0.21511999
#> [28,]  0.45004407
#> [29,]  0.65698660
#> [30,]  0.82308088
#> [31,]  0.93799998
#> [32,]  0.99459878
#> attr(,"class")
#> [1] "ts_data" "matrix"  "array"  
#> attr(,"sw")
#> [1] 1

#range of columns
data10[,1:2]
#>                t9          t8
#>  [1,]  0.00000000  0.24740396
#>  [2,]  0.24740396  0.47942554
#>  [3,]  0.47942554  0.68163876
#>  [4,]  0.68163876  0.84147098
#>  [5,]  0.84147098  0.94898462
#>  [6,]  0.94898462  0.99749499
#>  [7,]  0.99749499  0.98398595
#>  [8,]  0.98398595  0.90929743
#>  [9,]  0.90929743  0.77807320
#> [10,]  0.77807320  0.59847214
#> [11,]  0.59847214  0.38166099
#> [12,]  0.38166099  0.14112001
#> [13,]  0.14112001 -0.10819513
#> [14,] -0.10819513 -0.35078323
#> [15,] -0.35078323 -0.57156132
#> [16,] -0.57156132 -0.75680250
#> [17,] -0.75680250 -0.89498936
#> [18,] -0.89498936 -0.97753012
#> [19,] -0.97753012 -0.99929279
#> [20,] -0.99929279 -0.95892427
#> [21,] -0.95892427 -0.85893449
#> [22,] -0.85893449 -0.70554033
#> [23,] -0.70554033 -0.50827908
#> [24,] -0.50827908 -0.27941550
#> [25,] -0.27941550 -0.03317922
#> [26,] -0.03317922  0.21511999
#> [27,]  0.21511999  0.45004407
#> [28,]  0.45004407  0.65698660
#> [29,]  0.65698660  0.82308088
#> [30,]  0.82308088  0.93799998
#> [31,]  0.93799998  0.99459878
#> [32,]  0.99459878  0.98935825
#> attr(,"class")
#> [1] "ts_data" "matrix"  "array"  
#> attr(,"sw")
#> [1] 2

#range of rows and columns
data10[12:13,1:2]
#>            t9         t8
#> [1,] 0.381661  0.1411200
#> [2,] 0.141120 -0.1081951
#> attr(,"class")
#> [1] "ts_data" "matrix"  "array"  
#> attr(,"sw")
#> [1] 2

#single line and a range of columns
#'data10[12,1:2]

#range of lines and a single column
data10[12:13,1]
#>            t9
#> [1,] 0.381661
#> [2,] 0.141120
#> attr(,"class")
#> [1] "ts_data" "matrix"  "array"  
#> attr(,"sw")
#> [1] 1

#single observation
data10[12,1]
#>            t9
#> [1,] 0.381661
#> attr(,"class")
#> [1] "ts_data" "matrix"  "array"  
#> attr(,"sw")
#> [1] 1