k-fold training and test partition object

Splits a dataset into training and test sets based on k-fold cross-validation. The function takes a list of data partitions (folds) and a specified fold index k. It returns the data corresponding to the k-th fold as the test set, and combines all other folds to form the training set.

Usage

train_test_from_folds(folds, k)

Arguments

folds

data partitioned into folds

k

k-fold for test set, all reminder for training set

Value

returns a list with two elements:

train: A data frame containing the combined data from all folds except the k-th fold, used as the training set.

test: A data frame corresponding to the k-th fold, used as the test set.

Examples

# Create k-fold partitions of a dataset (e.g., iris)
folds <- k_fold(sample_random(), iris, k = 5)

# Use the first fold as the test set and combine the remaining folds for the training set
train_test_split <- train_test_from_folds(folds, k = 1)

# Display the training set
head(train_test_split$train)
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 48          4.6         3.2          1.4         0.2     setosa
#> 80          5.7         2.6          3.5         1.0 versicolor
#> 7           4.6         3.4          1.4         0.3     setosa
#> 81          5.5         2.4          3.8         1.1 versicolor
#> 18          5.1         3.5          1.4         0.3     setosa
#> 75          6.4         2.9          4.3         1.3 versicolor

# Display the test set
head(train_test_split$test)
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 30           4.7         3.2          1.6         0.2     setosa
#> 78           6.7         3.0          5.0         1.7 versicolor
#> 56           5.7         2.8          4.5         1.3 versicolor
#> 68           5.8         2.7          4.1         1.0 versicolor
#> 12           4.8         3.4          1.6         0.2     setosa
#> 134          6.3         2.8          5.1         1.5  virginica