ts_data

Time series data structure used in DAL Toolbox. It receives a vector (representing a time series) or a matrix y (representing a sliding windows). Internal ts_data is matrix of sliding windows with size sw. If sw equals to zero, it store a time series as a single matrix column.

Usage

ts_data(y, sw = 1)

Arguments

y

output variable

sw

integer: sliding window size.

Value

returns a ts_data object.

Examples

data(sin_data)
head(sin_data)
#>      x         y
#> 1 0.00 0.0000000
#> 2 0.25 0.2474040
#> 3 0.50 0.4794255
#> 4 0.75 0.6816388
#> 5 1.00 0.8414710
#> 6 1.25 0.9489846

data <- ts_data(sin_data$y)
ts_head(data)
#>             t0
#> [1,] 0.0000000
#> [2,] 0.2474040
#> [3,] 0.4794255
#> [4,] 0.6816388
#> [5,] 0.8414710
#> [6,] 0.9489846

data10 <- ts_data(sin_data$y, 10)
ts_head(data10)
#>             t9        t8        t7        t6        t5        t4        t3
#> [1,] 0.0000000 0.2474040 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950
#> [2,] 0.2474040 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859
#> [3,] 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859 0.9092974
#> [4,] 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859 0.9092974 0.7780732
#> [5,] 0.8414710 0.9489846 0.9974950 0.9839859 0.9092974 0.7780732 0.5984721
#> [6,] 0.9489846 0.9974950 0.9839859 0.9092974 0.7780732 0.5984721 0.3816610
#>             t2         t1         t0
#> [1,] 0.9839859  0.9092974  0.7780732
#> [2,] 0.9092974  0.7780732  0.5984721
#> [3,] 0.7780732  0.5984721  0.3816610
#> [4,] 0.5984721  0.3816610  0.1411200
#> [5,] 0.3816610  0.1411200 -0.1081951
#> [6,] 0.1411200 -0.1081951 -0.3507832