Time Series Adaptive Normalization (Exponential Moving Average - EMA)

Creates a normalization object for time series data using an Exponential Moving Average (EMA) method. This normalization approach adapts to changes in the time series and optionally removes outliers.

Usage

ts_norm_ean(remove_outliers = TRUE, nw = 0)

Arguments

remove_outliers

logical: if TRUE (default) outliers will be removed.

nw

windows size

Value

returns a ts_norm_ean object.

Examples

# time series to normalize
data(sin_data)

# convert to sliding windows
ts <- ts_data(sin_data$y, 10)
ts_head(ts, 3)
#>             t9        t8        t7        t6        t5        t4        t3
#> [1,] 0.0000000 0.2474040 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950
#> [2,] 0.2474040 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859
#> [3,] 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859 0.9092974
#>             t2        t1        t0
#> [1,] 0.9839859 0.9092974 0.7780732
#> [2,] 0.9092974 0.7780732 0.5984721
#> [3,] 0.7780732 0.5984721 0.3816610
summary(ts[,10])
#>        t0          
#>  Min.   :-0.99929  
#>  1st Qu.:-0.55091  
#>  Median : 0.05397  
#>  Mean   : 0.02988  
#>  3rd Qu.: 0.63279  
#>  Max.   : 0.99460  

# normalization
preproc <- ts_norm_ean()
preproc <- fit(preproc, ts)
tst <- transform(preproc, ts)
ts_head(tst, 3)
#>             t9        t8        t7        t6        t5        t4        t3
#> [1,] 0.1573972 0.2630526 0.3621389 0.4484953 0.5167527 0.5626671 0.5833837
#> [2,] 0.2718782 0.3709644 0.4573209 0.5255782 0.5714926 0.5922093 0.5864402
#> [3,] 0.4006181 0.4869745 0.5552319 0.6011462 0.6218629 0.6160938 0.5841976
#>             t2        t1        t0
#> [1,] 0.5776146 0.5457184 0.4896783
#> [2,] 0.5545440 0.4985038 0.4218041
#> [3,] 0.5281574 0.4514577 0.3588671
summary(tst[,10])
#>        t0        
#>  Min.   :0.2253  
#>  1st Qu.:0.2873  
#>  Median :0.4262  
#>  Mean   :0.4646  
#>  3rd Qu.:0.6407  
#>  Max.   :0.7764