Time Series Projection

Separates a ts_data object into input and output components for time series analysis. This function is useful for preparing data for modeling, where the input and output variables are extracted from a time series dataset.

Usage

ts_projection(ts)

Arguments

ts

matrix or data.frame containing the time series.

Value

returns a ts_projection object.

Examples

#setting up a ts_data
data(sin_data)
ts <- ts_data(sin_data$y, 10)

io <- ts_projection(ts)

#input data
ts_head(io$input)
#>             t9        t8        t7        t6        t5        t4        t3
#> [1,] 0.0000000 0.2474040 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950
#> [2,] 0.2474040 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859
#> [3,] 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859 0.9092974
#> [4,] 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859 0.9092974 0.7780732
#> [5,] 0.8414710 0.9489846 0.9974950 0.9839859 0.9092974 0.7780732 0.5984721
#> [6,] 0.9489846 0.9974950 0.9839859 0.9092974 0.7780732 0.5984721 0.3816610
#>             t2         t1
#> [1,] 0.9839859  0.9092974
#> [2,] 0.9092974  0.7780732
#> [3,] 0.7780732  0.5984721
#> [4,] 0.5984721  0.3816610
#> [5,] 0.3816610  0.1411200
#> [6,] 0.1411200 -0.1081951

#output data
ts_head(io$output)
#>              t0
#> [1,]  0.7780732
#> [2,]  0.5984721
#> [3,]  0.3816610
#> [4,]  0.1411200
#> [5,] -0.1081951
#> [6,] -0.3507832