Time Series Sample

Separates the ts_data into training and test. It separates the test size from the last observations minus an offset. The offset is important to allow replication under different recent origins. The data for train uses the number of rows of a ts_data minus the test size and offset.

Usage

ts_sample(ts, test_size = 1, offset = 0)

Arguments

ts

time series.

test_size

integer: size of test data (default = 1).

offset

integer: starting point (default = 0).

Value

returns a list with the two samples

Examples

#setting up a ts_data
data(sin_data)
ts <- ts_data(sin_data$y, 10)

#separating into train and test
test_size <- 3
samp <- ts_sample(ts, test_size)

#first five rows from training data
ts_head(samp$train, 5)
#>             t9        t8        t7        t6        t5        t4        t3
#> [1,] 0.0000000 0.2474040 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950
#> [2,] 0.2474040 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859
#> [3,] 0.4794255 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859 0.9092974
#> [4,] 0.6816388 0.8414710 0.9489846 0.9974950 0.9839859 0.9092974 0.7780732
#> [5,] 0.8414710 0.9489846 0.9974950 0.9839859 0.9092974 0.7780732 0.5984721
#>             t2        t1         t0
#> [1,] 0.9839859 0.9092974  0.7780732
#> [2,] 0.9092974 0.7780732  0.5984721
#> [3,] 0.7780732 0.5984721  0.3816610
#> [4,] 0.5984721 0.3816610  0.1411200
#> [5,] 0.3816610 0.1411200 -0.1081951

#last five rows from training data
ts_head(samp$train[-c(1:(nrow(samp$train)-5)),])
#>               t9          t8        t7        t6        t5        t4        t3
#> [1,] -0.27941550 -0.03317922 0.2151200 0.4500441 0.6569866 0.8230809 0.9380000
#> [2,] -0.03317922  0.21511999 0.4500441 0.6569866 0.8230809 0.9380000 0.9945988
#> [3,]  0.21511999  0.45004407 0.6569866 0.8230809 0.9380000 0.9945988 0.9893582
#> [4,]  0.45004407  0.65698660 0.8230809 0.9380000 0.9945988 0.9893582 0.9226042
#> [5,]  0.65698660  0.82308088 0.9380000 0.9945988 0.9893582 0.9226042 0.7984871
#>             t2        t1        t0
#> [1,] 0.9945988 0.9893582 0.9226042
#> [2,] 0.9893582 0.9226042 0.7984871
#> [3,] 0.9226042 0.7984871 0.6247240
#> [4,] 0.7984871 0.6247240 0.4121185
#> [5,] 0.6247240 0.4121185 0.1738895

#testing data
ts_head(samp$test)
#>             t9        t8        t7        t6        t5        t4        t3
#> [1,] 0.8230809 0.9380000 0.9945988 0.9893582 0.9226042 0.7984871 0.6247240
#> [2,] 0.9380000 0.9945988 0.9893582 0.9226042 0.7984871 0.6247240 0.4121185
#> [3,] 0.9945988 0.9893582 0.9226042 0.7984871 0.6247240 0.4121185 0.1738895
#>               t2          t1          t0
#> [1,]  0.41211849  0.17388949 -0.07515112
#> [2,]  0.17388949 -0.07515112 -0.31951919
#> [3,] -0.07515112 -0.31951919 -0.54402111